requirements-dev.txt
pytest==3.7.4
pytest-django==3.4.2